# trinity_config/openapi_trinity.yaml
openapi: 3.1.0
info:
  title: S25 Lumiere Trinity Bridge API
  description: |
    Complete OpenAPI schema for TRINITY Custom GPT → S25 Lumiere Cockpit.
    Updated for Phase 8 (Trinity full backend command — ops routes live).

    PRIMARY SERVER: https://api.smajor.org (Cloudflare tunnel → AlienStef :7777)
    SECONDARY: https://s25.smajor.org (same cockpit, alias)

    All GET endpoints are READ-ONLY PUBLIC and safe — they carry
    `x-openai-isConsequential: false` so ChatGPT Actions never prompt
    for consent on them. Only the POST/PUT endpoints that modify state
    require the X-S25-Secret header and may show a consent prompt.
  version: 8.0.0
  contact:
    name: Stef / S25 Lumiere
    url: https://smajor.org
servers:
  - url: https://api.smajor.org
    description: S25 API (PRIMARY — use this)
  - url: https://cockpit-alien.smajor.org
    description: Cockpit direct (same backend)
components:
  securitySchemes:
    S25Secret:
      type: apiKey
      in: header
      name: X-S25-Secret
      description: Shared secret for write endpoints
paths:
...
  /api/ops/agent/{name}/log:
    get:
      operationId: opsAgentLog
      summary: Tail last N lines of a specific agent log
      x-openai-isConsequential: false
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string, example: gemini_orchestrator }
        - name: n
          in: query
          required: false
          schema: { type: integer, default: 50, maximum: 500 }
      responses:
        '200': { description: Tail lines }
  /api/ops/agent/{name}/trigger:
    post:
      operationId: opsAgentTrigger
      summary: Run a cron agent immediately (one-shot)
      description: Triggers `python3 -m agents.{name}` synchronously with 90s timeout. Useful when Trinity wants fresh quant_brain or gemini_orchestrator output.
      security:
        - S25Secret: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Run result (stdout/stderr) }
  /api/ops/agent/{name}/restart:
    post:
      operationId: opsAgentRestart
      summary: Restart a systemd-user unit (e.g. cockpit)
      security:
        - S25Secret: []
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string, example: cockpit }
      responses:
        '200': { description: Restart result }
...
  /api/ops/git/pull:
    post:
      operationId: opsGitPull
      summary: Fast-forward pull from origin/main
      security:
        - S25Secret: []
      responses:
        '200': { description: Pull result }
  /api/ops/file:
    get:
      operationId: opsFileRead
      summary: Read a whitelisted infra file (agents/*, strategies/*, memory/*, trinity_config/*)
      description: Blocks .env, secrets, keys. Max 200KB.
      x-openai-isConsequential: false
      parameters:
        - name: path
          in: query
          required: true
          schema: { type: string, example: "agents/coinbase_executor.py" }
      responses:
        '200': { description: File content }
  /api/ops/dir:
    get:
      operationId: opsDirList
      summary: List whitelisted directory contents
      x-openai-isConsequential: false
      parameters:
        - name: path
          in: query
          required: false
          schema: { type: string, default: "agents" }
      responses:
        '200': { description: Directory listing }
...
  /api/ops/signal/block:
    post:
      operationId: opsBlockSource
      summary: Add a signal source to executor blocklist (stops ANY trade from that source)
      security:
        - S25Secret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string, example: ARKON5 }
      responses:
        '200': { description: Updated blocklist }
  /api/ops/signal/unblock:
    post:
      operationId: opsUnblockSource
      summary: Remove a source from executor blocklist
      security:
        - S25Secret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string }
      responses:
        '200': { description: Updated blocklist }
  /api/ops/kill-switch:
    get:
      operationId: opsKillSwitchGet
      summary: Is the emergency stop flag active?
      x-openai-isConsequential: false
      responses:
        '200': { description: Kill-switch state }
    post:
      operationId: opsKillSwitchSet
      summary: Activate or deactivate the emergency stop (trades fully blocked while active)
      description: Activate does NOT require the secret (safety). Deactivate requires X-S25-Secret.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                activate: { type: boolean, default: true }
                reason: { type: string }
      responses:
        '200': { description: New state }